requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.2
h11==0.14.0
idna==3.10
mysql-connector-python==9.0.0
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
starlette==0.39.2
typing_extensions==4.12.2
uvicorn==0.31.1
SQLAlchemy==2.0.32
python-dotenv
azure-storage-blob==12.23.1